backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==1.4.36
pymysql==1.0.2
python-dotenv==1.0.0
